requirements.txt
python-dotenv
pandas
numpy
certifi
# pymongo[srv]
# pymongo
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart



# -e .